+++ services/evm/rotation_signer_test.go
@@ -1,13 +1,14 @@
-package requester
+package evm
 
 import (
 	"fmt"
+	"sync"
+	"testing"
+
 	"github.com/onflow/flow-go-sdk/crypto"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/exp/maps"
-	"sync"
-	"testing"
 )
 
 func Test_KeyRotation(t *testing.T) {
